# frontend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Login API",
    "version": "1.0.0",
    "description": "Login system APIs"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "summary": "Login to your account",
        "tags": [
          "auth-controller"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "You are logged",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "406": {
            "description": "Wrong email or password"
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Create a new account",
        "tags": [
          "auth-controller"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user was successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterRequest"
                }
              }
            }
          },
          "500": {
            "description": "Some server error"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Return the list of all users",
        "parameters": [
          {
            "in": "header",
            "name": "token",
            "schema": {
              "type": "string",
              "description": "The user token"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "description": "The current page number"
            }
          }
        ],
        "tags": [
          "user-controller"
        ],
        "responses": {
          "200": {
            "description": "The list of the users",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableDataDto"
                }
              }
            }
          }
        }
      }
    },
    "/user/current": {
      "get": {
        "summary": "Return the curent user data.",
        "parameters": [
          {
            "in": "header",
            "name": "token",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "tags": [
          "user-controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update user data",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          }
        },
        "tags": [
          "user-controller"
        ],
        "responses": {
          "200": {
            "description": "The user data has been updated successfully"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserDto": {
        "type": "object",
        "required": [
          "lastName",
          "firstName",
          "email",
          "city",
          "state",
          "country",
          "profilePictureUrl"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "minLength": 1,
            "description": "The unique id of each user",
            "readOnly": true
          },
          "firstName": {
            "type": "string",
            "minLength": 1,
            "description": "The first name of each user"
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "description": "The last name of each user"
          },
          "email": {
            "type": "string",
            "minLength": 1,
            "description": "The email of each user"
          },
          "city": {
            "type": "string",
            "minLength": 1,
            "description": "The city of each user"
          },
          "state": {
            "type": "string",
            "minLength": 1,
            "description": "The gender for each user"
          },
          "country": {
            "type": "string",
            "minLength": 1,
            "description": "The country for each user"
          },
          "profilePictureUrl": {
            "type": "string",
            "minLength": 1,
            "description": "The profile picture URL"
          },
          "userRole": {
            "$ref": "#/components/schemas/RoleDto"
          }
        },
        "example": {
          "id": 0,
          "firstName": "string",
          "lastName": "string",
          "email": "string",
          "city": "string",
          "state": "string",
          "country": "string",
          "profilePictureUrl": "string",
          "userRole": {
            "id": 0,
            "name": "ADMIN"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "example": {
          "email": "string",
          "password": "string"
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "firstName",
          "lastName",
          "email",
          "password"
        ],
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "example": {
          "firstName": "string",
          "lastName": "string",
          "email": "string",
          "password": "string"
        }
      },
      "RoleDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "enum": [
              "ADMIN",
              "SUPERADMIN",
              "USER"
            ]
          }
        }
      },
      "TableDataDto": {
        "required": [
          "page"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "minLength": 1
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataDto"
            }
          }
        },
        "example": {
          "page": 1,
          "data": {
            "id": 0,
            "firstName": "string",
            "lastName": "string",
            "email": "string",
            "created_at": "string",
            "userRole": {
              "id": 0,
              "role": "SUPERADMIN"
            }
          }
        }
      },
      "DataDto": {
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "minLength": 1
          },
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "created_at": {
            "type": "string",
            "minLength": 1
          },
          "userRole": {
            "$ref": "#/components/schemas/RoleDto"
          }
        },
        "example": {
          "id": 0,
          "firstName": "string",
          "lastName": "string",
          "email": "string",
          "created_at": "string",
          "userRole": {
            "id": 0,
            "role": "SUPERADMIN"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "auth-controller"
    },
    {
      "name": "user-controller"
    }
  ]
}